+++ expansion/src/test/java/no/unit/nva/expansion/ResourceExpansionServiceTest.java
@@ -12,6 +12,7 @@ import static no.unit.nva.model.testing.associatedartifacts.AssociatedArtifactsG
 import static no.unit.nva.publication.PublicationServiceConfig.API_HOST;
 import static no.unit.nva.publication.testing.http.RandomPersonServiceResponse.randomUri;
 import static no.unit.nva.testutils.RandomDataGenerator.randomElement;
+import static no.unit.nva.testutils.RandomDataGenerator.randomInteger;
 import static no.unit.nva.testutils.RandomDataGenerator.randomString;
 import static nva.commons.core.ioutils.IoUtils.stringFromResources;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -67,6 +68,12 @@ import no.unit.nva.model.associatedartifacts.RelationType;
 import no.unit.nva.model.associatedartifacts.file.File;
 import no.unit.nva.model.associatedartifacts.file.PendingFile;
 import no.unit.nva.model.associatedartifacts.file.PendingOpenFile;
+import no.unit.nva.model.instancetypes.degree.ArtisticDegreePhd;
+import no.unit.nva.model.instancetypes.degree.DegreeBachelor;
+import no.unit.nva.model.instancetypes.degree.DegreeLicentiate;
+import no.unit.nva.model.instancetypes.degree.DegreeMaster;
+import no.unit.nva.model.instancetypes.degree.DegreePhd;
+import no.unit.nva.model.instancetypes.degree.OtherStudentWork;
 import no.unit.nva.model.instancetypes.journal.AcademicArticle;
 import no.unit.nva.model.role.Role;
 import no.unit.nva.model.role.RoleType;
@@ -78,6 +85,7 @@ import no.unit.nva.publication.model.business.GeneralSupportRequest;
 import no.unit.nva.publication.model.business.Message;
 import no.unit.nva.publication.model.business.MessageStatus;
 import no.unit.nva.publication.model.business.PublishingRequestCase;
+import no.unit.nva.publication.model.business.PublishingWorkflow;
 import no.unit.nva.publication.model.business.Resource;
 import no.unit.nva.publication.model.business.TicketEntry;
 import no.unit.nva.publication.model.business.TicketStatus;
@@ -119,7 +127,7 @@ class ResourceExpansionServiceTest extends ResourcesLocalTest {
     private static final String VIEWED_BY = "viewedBy";
     private static final String WORKFLOW = "workflow";
     private static final String ASSIGNEE = "assignee";
-    private static final String OWNERAFFILIATION = "ownerAffiliation";
+    private static final String OWNER_AFFILIATION = "ownerAffiliation";
     private static final String FINALIZED_BY = "finalizedBy";
     public static final String APPROVED_FILES = "approvedFiles";
     public static final String FILES_FOR_APPROVAL = "filesForApproval";
@@ -193,7 +201,7 @@ class ResourceExpansionServiceTest extends ResourcesLocalTest {
 
         assertThat(ticket,
                    doesNotHaveEmptyValuesIgnoringFields(Set.of(WORKFLOW, ASSIGNEE, FINALIZED_BY,
-                                                               FINALIZED_DATE, OWNERAFFILIATION, APPROVED_FILES,
+                                                               FINALIZED_DATE, OWNER_AFFILIATION, APPROVED_FILES,
                                                                FILES_FOR_APPROVAL,
                                                                RESPONSIBILITY_AREA, TICKET_EVENT, VIEWED_BY)));
         var expectedPublicationId = constructExpectedPublicationId(publication);
@@ -804,12 +812,10 @@ class ResourceExpansionServiceTest extends ResourcesLocalTest {
 
     private TicketEntry persistPublishingRequestContainingExistingUnpublishedFiles(Publication publication)
         throws ApiGatewayException {
-        var publishingRequest = (PublishingRequestCase) PublishingRequestCase.createNewTicket(publication, PublishingRequestCase.class,
-                                                                                              SortableIdentifier::next)
-                                                            .withOwner(UserInstance.fromPublication(publication).getUsername())
-                                                            .withOwnerAffiliation(publication.getResourceOwner().getOwnerAffiliation());
-        publishingRequest.withFilesForApproval(TicketTestUtils.getFilesForApproval(publication));
-        return publishingRequest.persistNewTicket(ticketService);
+        return PublishingRequestCase.create(Resource.fromPublication(publication),
+                                            UserInstance.fromPublication(publication),
+                                            PublishingWorkflow.REGISTRATOR_PUBLISHES_METADATA_ONLY)
+                   .persistNewTicket(ticketService);
     }
 
     private static List<Contributor> extractContributorsWithId(URI id, Publication publication) {
@@ -980,7 +986,9 @@ class ResourceExpansionServiceTest extends ResourcesLocalTest {
     }
 
     private Publication persistDraftPublicationWithoutDoi() throws BadRequestException {
-        var publication = randomPublication().copy().withDoi(null).withStatus(DRAFT).build();
+        var degrees = List.of(DegreeBachelor.class, DegreeMaster.class, DegreePhd.class,
+                              ArtisticDegreePhd.class, DegreeLicentiate.class, OtherStudentWork.class);
+        var publication = randomPublication(degrees.get(randomInteger(degrees.size()))).copy().withDoi(null).withStatus(DRAFT).build();
         return Resource.fromPublication(publication)
                    .persistNew(resourceService, UserInstance.fromPublication(publication));
     }
@@ -999,6 +1007,7 @@ class ResourceExpansionServiceTest extends ResourcesLocalTest {
         publishingRequest.setApprovedFiles(expandedPublishingRequest.getApprovedFiles());
         publishingRequest.setFilesForApproval(expandedPublishingRequest.getFilesForApproval());
         publishingRequest.setOwnerAffiliation(expandedPublishingRequest.getOrganization().id());
+        publishingRequest.setWorkflow(expandedPublishingRequest.getWorkflow());
         return publishingRequest;
     }
