+++ publication-commons/src/main/java/no/unit/nva/publication/model/business/DoiRequest.java
@@ -1,14 +1,6 @@
 package no.unit.nva.publication.model.business;
 
 import static java.util.Objects.nonNull;
-import static no.unit.nva.publication.model.business.TicketEntry.Constants.ASSIGNEE_FIELD;
-import static no.unit.nva.publication.model.business.TicketEntry.Constants.CREATED_DATE_FIELD;
-import static no.unit.nva.publication.model.business.TicketEntry.Constants.CUSTOMER_ID_FIELD;
-import static no.unit.nva.publication.model.business.TicketEntry.Constants.IDENTIFIER_FIELD;
-import static no.unit.nva.publication.model.business.TicketEntry.Constants.OWNER_AFFILIATION_FIELD;
-import static no.unit.nva.publication.model.business.TicketEntry.Constants.STATUS_FIELD;
-import static nva.commons.core.attempt.Try.attempt;
-import com.fasterxml.jackson.annotation.JsonAlias;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonTypeInfo.Id;
@@ -17,7 +9,6 @@ import java.time.Clock;
 import java.time.Instant;
 import java.util.Collections;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
 import no.unit.nva.identifiers.SortableIdentifier;
 import no.unit.nva.model.Publication;
@@ -29,12 +20,10 @@ import no.unit.nva.publication.model.business.publicationstate.DoiRequestedEvent
 import no.unit.nva.publication.model.business.publicationstate.TicketEvent;
 import no.unit.nva.publication.model.storage.DoiRequestDao;
 import no.unit.nva.publication.model.storage.TicketDao;
-import no.unit.nva.publication.service.impl.ResourceService;
 import no.unit.nva.publication.storage.model.exceptions.IllegalDoiRequestUpdate;
 import nva.commons.apigateway.exceptions.ApiGatewayException;
 import nva.commons.apigateway.exceptions.ConflictException;
 import nva.commons.core.JacocoGenerated;
-import nva.commons.core.attempt.Try;
 
 @JsonTypeInfo(use = Id.NAME, property = "type")
 @SuppressWarnings({"PMD.GodClass", "PMD.ExcessivePublicCount", "PMD.TooManyFields"})
@@ -51,23 +40,8 @@ public class DoiRequest extends TicketEntry {
                PublicationStatus.PUBLISHED_METADATA,
                PublicationStatus.DRAFT);
     public static final String DOI_REQUEST_APPROVAL_FAILURE = "Cannot approve DoiRequest for non-published publication";
-    @JsonProperty(IDENTIFIER_FIELD)
-    private SortableIdentifier identifier;
-    @JsonProperty(STATUS_FIELD)
-    private TicketStatus status;
     @JsonProperty(RESOURCE_STATUS_FIELD)
     private PublicationStatus resourceStatus;
-    @JsonProperty
-    private Instant modifiedDate;
-    @JsonProperty(CREATED_DATE_FIELD)
-    @JsonAlias("date")
-    private Instant createdDate;
-    @JsonProperty(CUSTOMER_ID_FIELD)
-    private URI customerId;
-    @JsonProperty(ASSIGNEE_FIELD)
-    private Username assignee;
-    @JsonProperty(OWNER_AFFILIATION_FIELD)
-    private URI ownerAffiliation;
     @JsonProperty("ticketEvent")
     private TicketEvent ticketEvent;
 
@@ -108,55 +82,8 @@ public class DoiRequest extends TicketEntry {
     }
 
     @Override
-    public SortableIdentifier getIdentifier() {
-        return identifier;
-    }
-
-    @Override
-    public void setIdentifier(SortableIdentifier identifier) {
-        this.identifier = identifier;
-    }
-
-    @Override
-    public Publication toPublication(ResourceService resourceService) {
-        return Optional.ofNullable(getResourceIdentifier())
-                   .map(attempt(resourceService::getPublicationByIdentifier))
-                   .map(Try::orElseThrow)
-                   .orElse(null);
-    }
-
-    @Override
     public String getType() {
         return DoiRequest.TYPE;
-    }
-
-    @Override
-    public Instant getCreatedDate() {
-        return createdDate;
-    }
-
-    @Override
-    public void setCreatedDate(Instant createdDate) {
-        this.createdDate = createdDate;
-    }
-
-    @Override
-    public Instant getModifiedDate() {
-        return modifiedDate;
-    }
-
-    @Override
-    public void setModifiedDate(Instant modifiedDate) {
-        this.modifiedDate = modifiedDate;
-    }
-
-    @Override
-    public URI getCustomerId() {
-        return customerId;
-    }
-
-    public void setCustomerId(URI customerId) {
-        this.customerId = customerId;
     }
 
     @Override
@@ -218,46 +145,12 @@ public class DoiRequest extends TicketEntry {
                    .build();
     }
 
-    @Override
-    public TicketStatus getStatus() {
-        return status;
-    }
-
-    @Override
-    public void setStatus(TicketStatus status) {
-        this.status = status;
-    }
-
-    @Override
-    public Username getAssignee() {
-        return assignee;
-    }
-
-    @Override
-    public void setAssignee(Username assignee) {
-        this.assignee = assignee;
-    }
-
-    @Override
-    public URI getOwnerAffiliation() {
-        return ownerAffiliation;
-    }
-
-    @Override
-    public void setOwnerAffiliation(URI ownerAffiliation) {
-        this.ownerAffiliation = ownerAffiliation;
-    }
-
     public TicketEvent getTicketEvent() {
         return ticketEvent;
     }
 
     public void setTicketEvent(TicketEvent ticketEvent) {
         this.ticketEvent = ticketEvent;
-    }
-
-    @Override
-    public void validateAssigneeRequirements(Publication publication) {
     }
 
     public DoiRequest update(Resource resource) {
