+++ publication-commons/src/test/java/no/unit/nva/publication/model/storage/DaoTest.java
@@ -16,6 +16,7 @@ import static no.unit.nva.publication.storage.model.DatabaseConstants.PRIMARY_KE
 import static no.unit.nva.publication.storage.model.DatabaseConstants.RESOURCES_TABLE_NAME;
 import static no.unit.nva.publication.storage.model.DatabaseConstants.STATUS_INDEX_FIELD_PREFIX;
 import static no.unit.nva.testutils.RandomDataGenerator.randomDoi;
+import static no.unit.nva.testutils.RandomDataGenerator.randomInteger;
 import static no.unit.nva.testutils.RandomDataGenerator.randomString;
 import static no.unit.nva.testutils.RandomDataGenerator.randomUri;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -55,7 +56,12 @@ import no.unit.nva.model.ResourceOwner;
 import no.unit.nva.model.Username;
 import no.unit.nva.model.additionalidentifiers.AdditionalIdentifier;
 import no.unit.nva.model.funding.FundingBuilder;
+import no.unit.nva.model.instancetypes.degree.ArtisticDegreePhd;
 import no.unit.nva.model.instancetypes.degree.DegreeBachelor;
+import no.unit.nva.model.instancetypes.degree.DegreeLicentiate;
+import no.unit.nva.model.instancetypes.degree.DegreeMaster;
+import no.unit.nva.model.instancetypes.degree.DegreePhd;
+import no.unit.nva.model.instancetypes.degree.OtherStudentWork;
 import no.unit.nva.model.role.Role;
 import no.unit.nva.model.role.RoleType;
 import no.unit.nva.publication.model.business.DoiRequest;
@@ -130,7 +136,9 @@ class DaoTest extends ResourcesLocalTest {
     }
 
     public static Publication draftDegreePublicationWithoutDoi() {
-        return randomPublication(DegreeBachelor.class).copy()
+        var degrees = List.of(DegreeBachelor.class, DegreeMaster.class, DegreePhd.class,
+                              ArtisticDegreePhd.class,DegreeLicentiate.class, OtherStudentWork.class);
+        return randomPublication(degrees.get(randomInteger(degrees.size()))).copy()
                    .withStatus(PublicationStatus.DRAFT)
                    .withDoi(null)
                    .build();
